Correct rank by I-st ranking with more data
With more data there are deviations, e.g. for some classes I-st ranking is
lower than II-nd ranking or vice versa. This leads to incorrect rank, so
find the rank in the list sorted by I-st rank and display it for reference.
The table remains ordered by II-nd ranking, so II-nd rank is still OK.